Convert to Inertia views
