+++ resources/js/addon.js
@@ -1,9 +1,9 @@
 import MuxMirrorFieldtype from './components/MuxMirrorFieldtype.vue'
 import MuxMirrorIndexFieldtype from './components/MuxMirrorIndexFieldtype.vue'
-import MuxMirroredListing from './components/MuxMirroredListing.vue'
-import MuxLibraryListing from './components/MuxLibraryListing.vue'
+import MirroredAssetsPage from './pages/MirroredAssets.vue'
+import MuxLibraryPage from './pages/MuxLibrary.vue'
 
 Statamic.$components.register('mux_mirror-fieldtype', MuxMirrorFieldtype)
 Statamic.$components.register('mux_mirror-fieldtype-index', MuxMirrorIndexFieldtype)
-Statamic.$components.register('mux-mirrored-listing', MuxMirroredListing)
-Statamic.$components.register('mux-library-listing', MuxLibraryListing)
+Statamic.$inertia.register('MuxAssetsPage', MirroredAssetsPage)
+Statamic.$inertia.register('MuxLibraryPage', MuxLibraryPage)
